Fixed add product, finished

+++ src/hooks/useProduct.ts
@@ -69,7 +69,6 @@ const useProduct = () => {
         const [errors, productDto] = await ProductDTO.created({ ...data });
         if (errors) throw new Error(errors);
         const dataNew = (await dispatch(productAsyncThunks.onSaveProduct(productDto!))).payload;
-        console.log(dataNew);
         Validator.httpValidation(dataNew as any);
         return ProductMapper.getItem(dataNew);
       }
@@ -97,25 +96,47 @@ const useProduct = () => {
     }
   });
 
+  const useProductIsEcommerceMutation = (pagination?: IPagination) => useMutation<ProductEntity, Error, ProductEntity>({
+    mutationFn: async (data) => { 
+      const [errors, productDto] = await ProductDTO.updated({ ...data });
+      if (errors) throw new Error(errors);
+      const dataNew = (await dispatch(productAsyncThunks.onEditProductIsEcommerce(productDto!))).payload;
+      Validator.httpValidation(dataNew as any);
+      if (dataNew) {
+        return {...data, isEcommerce: !data.isEcommerce };
+      }
+      return data;
+    },
+    onSuccess: (data) => {
+      pagination && queryClient.invalidateQueries({
+        queryKey: ['product-list', { ...pagination }]
+      });
+      return data;
+    }
+  });
+
   const useProductEliminateItemsMutation = (contents: ContentsDelete[]) => useMutation<boolean, Error, ProductEntity>({
     mutationFn: async (data) => {
-      const [errors, productDto] = await ProductDTO.created({ ...data });
-      if (errors) throw new Error(errors);
-      if (contents.includes('colors')) {
-        const values = (await dispatch(productAsyncThunks.onDeleteProductColors(productDto!))).payload;
-        Validator.httpValidation(values as any);
-      }
-      if (contents.includes('images')) {
-        const values = (await dispatch(productAsyncThunks.onDeleteProductImages(productDto!))).payload;
-        Validator.httpValidation(values as any);
-      }
-      if (contents.includes('dimensions')) {
-        const values = (await dispatch(productAsyncThunks.onDeleteProductDimensions(productDto!))).payload;
-        Validator.httpValidation(values as any);
-      }
-      if (contents.includes('details')) {
-        const values = (await dispatch(productAsyncThunks.onDeleteProductDetails(productDto!))).payload;
-        Validator.httpValidation(values as any);
+      if (data.productId) {
+        if (data.productId <= 0) return true;
+        const [errors, productDto] = await ProductDTO.updated({ ...data });
+        if (errors) throw new Error(errors);
+        if (contents.includes('colors')) {
+          const values = (await dispatch(productAsyncThunks.onDeleteProductColors(productDto!))).payload;
+          Validator.httpValidation(values as any);
+        }
+        if (contents.includes('images')) {
+          const values = (await dispatch(productAsyncThunks.onDeleteProductImages(productDto!))).payload;
+          Validator.httpValidation(values as any);
+        }
+        if (contents.includes('dimensions')) {
+          const values = (await dispatch(productAsyncThunks.onDeleteProductDimensions(productDto!))).payload;
+          Validator.httpValidation(values as any);
+        }
+        if (contents.includes('details')) {
+          const values = (await dispatch(productAsyncThunks.onDeleteProductDetails(productDto!))).payload;
+          Validator.httpValidation(values as any);
+        }
       }
       return true;
     },
@@ -134,6 +155,7 @@ const useProduct = () => {
     useProductListQuery,
     useProductQuery,
     useProductMutation,
+    useProductIsEcommerceMutation,
     useProductEliminateItemsMutation,
     filterProductExistByName
   };
